requirements.txt
python-binance==0.7.5
numpy==1.19.3
pandas==1.1.4
ta==0.6.0